fixing solr config path, populating solr with assets
Having major problems getting indexing to work on Play startup, in dev mode
calling server.add causes Play to reload the application for some reason.  For
now I got around it by adding a simple controller action for performing the
indexing.  This happens when using either the embedded server or an external
server via http.  This problem doesn't seem to occur when collins is run in
production mode.