Introducing state attribute for asset
A state represents the current operational state of an asset. The state may be
status specific (e.g. 'Hardware Problem' is associated only with the Maintenance
status), or available for any state (e.g. Running). While a status represents
the lifecycle of an asset, the state represents the discrete states within a
status. For instance, a machine that is Allocated may be Running or Terminated
but in both cases still allocated.

The primary driver for this change is to support discrete states within
Maintenance, so that techs can determine easily why a machine is in maintenance.
Additionally being able to hook ticket creation into changing the maintenance
state is ideal. For instance if something is having a hardware problem you may
want to create a hardware ticket in the datacenter queue, but if the asset is
having a network problem you may want to create a ticket in the network queue.

Still left TODO:

  * Unit tests for state
  * 'Thin' JSON representation for a state, for use when querying an asset
  * API support for adding/listing/managing states
  * Configuration for default state of an asset when coming out of maintenance
  * Support for managing asset state via the API
  * UI support for managing asset state
  * Callback support for matching on both a status and a state