Graph plugin support refs #41
Added support for showing graphs in the asset details view. Currently only
support for fibr graphs exist which won't generally be useful to anyone outside
of Tumblr. Creating support for a different graph system is pretty trivial so
hopefully there will be some community support.

Need to only show the tab pane if the plugin is enabled. Also make sure that the
config is reloadable.