collins-shell threads support
You can use collins-shell asset find --exec='cmd' but this is slow if you need
to perform the same operation against a number of hosts. You can now specify
--threads=N to run the command in N threads.

The only fix we have is to override say_status from thor to synchronize so that
output of concurrent execs doesn't smash each other.